Remade Controller and RestController, jsp pages, js files, validator, dao, ExceptionHandler for Specialty.class

+++ mvc-war/src/main/java/com/arttu/mvc/controller/SpecialtyController.java
@@ -1,21 +1,23 @@
 package com.arttu.mvc.controller;
 
+import com.arttu.mvc.exception.department.DepartmentNotFoundException;
+import com.arttu.mvc.exception.specialty.SpecialtyNotFoundException;
+import com.arttu.mvc.model.Department;
+import com.arttu.mvc.model.Item;
 import com.arttu.mvc.model.Specialty;
 import com.arttu.mvc.service.DepartmentService;
 import com.arttu.mvc.service.SpecialtyService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
-@RestController
+@Controller
 public class SpecialtyController {
 
     private static final Logger LOGGER = LogManager.getLogger(SpecialtyController.class);
@@ -35,72 +37,68 @@ public class SpecialtyController {
     @GetMapping("/specialty")
     public ModelAndView department() {
         ModelAndView modelAndView = new ModelAndView("specialty/specialty");
-        modelAndView.addObject("specialtyList", specialtyService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
+        List<Specialty> specialties = specialtyService.findAll();
+        if (specialties == null) {
+            LOGGER.error("department() method of SpecailtyController class");
+            throw new SpecialtyNotFoundException("Specialty not found");
+        }
+        modelAndView.addObject("specialtyList", specialties);
         return modelAndView;
     }
 
     @GetMapping("/specialty/find/{id}")
     public ModelAndView findById(@PathVariable int id) {
         ModelAndView modelAndView = new ModelAndView("specialty/specialty");
-        modelAndView.addObject("specialty", specialtyService.findById(id));
-        modelAndView.addObject("itemList", specialtyService.findAttachedItems(id));
-        modelAndView.setStatus(HttpStatus.OK);
+        Specialty specialty = specialtyService.findById(id);
+        List<Item> itemList = specialtyService.findAttachedItems(id);
+        if (specialty == null) {
+            LOGGER.error("findById method of SpecialtyController class");
+            throw new SpecialtyNotFoundException("Specialty not found");
+        }
+        if (itemList == null) {
+            LOGGER.error("findById method of SpecialtyController class");
+            throw new SpecialtyNotFoundException("Item not found");
+        }
+        modelAndView.addObject("specialty", specialty);
+        modelAndView.addObject("itemList", itemList);
         return modelAndView;
     }
 
     @GetMapping("/specialty/add")
     public ModelAndView add() {
         ModelAndView modelAndView = new ModelAndView("specialty/addSpecialty");
-        modelAndView.addObject("departmentList", departmentService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
+        List<Department> departments = departmentService.findAll();
+        if (departments == null) throw new DepartmentNotFoundException("Department not found");
+        modelAndView.addObject("departmentList", departments);
         return modelAndView;
-    }
-
-    @PostMapping(value = "/specialty/add", consumes = {MediaType.APPLICATION_JSON_VALUE},
-            produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<Map<String, Object>> add(@RequestBody Specialty specialty) {
-        Map<String, Object> response = new HashMap<>();
-        HttpStatus status = HttpStatus.OK;
-        try {
-            specialtyService.add(specialty);
-            response.put("stat", 1);
-        } catch (Exception e) {
-            LOGGER.error(e);
-            response.put("stat", 0);
-            status = HttpStatus.BAD_REQUEST;
-        }
-        return new ResponseEntity<>(response, status);
     }
 
     @GetMapping(value = "/specialty/edit/{id}")
     public ModelAndView edit(@PathVariable int id) {
         ModelAndView modelAndView = new ModelAndView("specialty/editSpecialty");
-        modelAndView.addObject("specialtyList", specialtyService.findById(id));
-        modelAndView.addObject("departmentList", departmentService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
-        return modelAndView;
-    }
-
-    @PostMapping(value = "/specialty/edit", consumes = {MediaType.APPLICATION_JSON_VALUE},
-            produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<Map<String, Object>> edit(@RequestBody Specialty specialty) {
-        Map<String, Object> response = new HashMap<>();
-        HttpStatus status = HttpStatus.OK;
-        try {
-            specialtyService.edit(specialty);
-            response.put("stat", 1);
-        } catch (Exception e) {
-            LOGGER.error(e);
-            response.put("stat", 0);
-            status = HttpStatus.BAD_REQUEST;
+        Specialty specialty = specialtyService.findById(id);
+        List<Department> departments = departmentService.findAll();
+        if (specialty == null) {
+            LOGGER.error("Edit method of SpecialtyController class");
+            throw new SpecialtyNotFoundException("Specialty not found");
         }
-        return new ResponseEntity<>(response, status);
+        if (departments == null) {
+            LOGGER.error("Edit method of SpecialtyController class");
+            throw new DepartmentNotFoundException("Department not found");
+        }
+        modelAndView.addObject("specialtyList", specialty);
+        modelAndView.addObject("departmentList", departments);
+        return modelAndView;
     }
 
     @GetMapping(value = "/specialty/delete/{id}")
     public ModelAndView delete(@PathVariable int id) {
-        specialtyService.deleteById(id);
+        Specialty specialty = specialtyService.findById(id);
+        if (specialty == null) {
+            throw new SpecialtyNotFoundException("Specialty not found");
+        } else {
+            specialtyService.deleteById(id);
+        }
         return new ModelAndView("redirect:/specialty");
     }
 }
